+++ docstrfmt/main.py
@@ -21,7 +21,11 @@ from typing import TYPE_CHECKING, Any
 
 import click
 import libcst as cst
-import toml
+
+if sys.version_info >= (3, 11):
+    import tomllib as toml  # pragma: no cover
+else:
+    import toml  # pragma: no cover
 from black import (
     DEFAULT_LINE_LENGTH,
     Mode,
@@ -122,7 +126,11 @@ def _parse_pyproject_config(
         value = pyproject_toml if pyproject_toml else None
     if value:
         try:
-            pyproject_toml = toml.load(value)
+            if sys.version_info >= (3, 11):
+                with open(value, "rb") as f:  # pragma: no cover
+                    pyproject_toml = toml.load(f)  # pragma: no cover
+            else:
+                pyproject_toml = toml.load(value)  # pragma: no cover
             config = pyproject_toml.get("tool", {}).get("docstrfmt", {})
             config = {
                 k.replace("--", "").replace("-", "_"): v for k, v in config.items()